% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_sd.R
\name{beta_sd}
\alias{beta_sd}
\title{Compute standard deviation of Beta distribution.}
\usage{
beta_sd(alpha, beta, a, b)
}
\arguments{
\item{alpha, beta}{Integers. Need to be greater than zero. Control shape of distribution. \verb{alpha = 4, beta = 4} gives a bell-shaped distribution
\verb{alpha = 1, beta = 1} gives a Uniform distribution. \verb{alpha = 0.5, beta = 0.5} gives
a high-variance distribution where very low or very high values or most likely.}

\item{a, b}{Integers. Lower and upper bound of unit-level causal effects.}
}
\value{
An integer.
}
\description{
Compute standard deviation of Beta distribution.
}
\details{
Assuming that unit-level causal effects vary between \code{a} and \code{b}, this function
computes implied standard deviations of the distribution of causal effects,
assuming a Beta distribution with parameters \code{alpha} and \code{beta}.

The computed standard deviation can be used to inform the magnitude of the
sensitivity parameter \code{sigma} in \code{\link[=post_instrument_bounds]{post_instrument_bounds()}}.
}
\examples{
x <- seq(0, 1, length.out = 100)
plot(x, dbeta(x, shape1 = 1, shape2 = 1), type = "l",
ylim = c(0, 3.5),
xlab = expression(paste(gamma[i])),
ylab = "Density")
lines(x, dbeta(x, shape1 = 0.5, shape2 = 0.5), type = "l",
     lty = "dashed")
lines(x, dbeta(x, shape1 = 4, shape2 = 4), type = "l",
     lty = "dotted")
legend(x = "top", c("Low Variance", "Medium Variance", "High Variance"),
      lty = c("dotted", "solid", "dashed"))

beta_sd(alpha = 1, beta = 1, a= 0, b = 1) # medium variance
beta_sd(alpha = 0.5, beta = 0.5, a = 0, b = 1) # high variance
beta_sd(alpha = 4, beta = 4, a = 0, b = 1) # low variance
}
\references{
Schuessler, J., Glynn, A. N., & Rueda, M.R. (2023). Post-Instrument Bias.
}
