% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carnegie_marinov}
\alias{carnegie_marinov}
\title{Data from Carnegie/Marinov 2017}
\format{
An object of class \code{data.frame} with 1767 rows and 6 columns.
}
\source{
\url{https://doi.org/10.15139/S3/X5WXLM}
}
\usage{
carnegie_marinov
}
\description{
A subset of data from Allison Carnegie and Nikolay Marinov (2017), "Foreign Aid, Human Rights, and Democracy Promotion:
Evidence from a Natural Experiment", American Journal of Political Science 61(3).
}
\keyword{datasets}
