% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_sensitivity.R
\name{estimate_sensitivity}
\alias{estimate_sensitivity}
\title{Estimate components for sensitivity analysis.}
\usage{
estimate_sensitivity(
  df,
  R = 1501,
  outcome,
  treat,
  inst,
  post.inst,
  x = NULL,
  fe = NULL,
  weights = NULL,
  seed,
  parallel = "no",
  ncpus = getOption("boot.ncpus", 1L),
  cl = NULL
)
}
\arguments{
\item{df}{Data frame}

\item{R}{Number of bootstrap replications}

\item{outcome}{Name of dependent variables of interest \code{Y}, a string}

\item{treat}{Name of treatment of interest \code{D}, a string}

\item{inst}{Name of instrument \code{Z}, a string}

\item{post.inst}{Name of post-instrument variable \code{M}, a string}

\item{x}{Names of controls, a string or vector of strings}

\item{fe}{Names of variables to be passed as fixed effects (or factors),
a string or vector of strings}

\item{weights}{Vector of weights to be used in all regressions}

\item{seed}{Seed to make bootstrap analysis replicable. Must be set to a number}

\item{parallel}{One of \code{c("no", "snow", "multicore")} to parallelize the bootstrap. Default is
\code{"no"}.}

\item{ncpus}{Integer: number of processes to be used in parallel operation}

\item{cl}{An optional parallel or snow cluster for use if \code{parallel = "snow"}.}
}
\value{
An object of class \code{boot}, where \code{t0} contains point estimates for all elements of the bounds estimator, and \code{t} contains R bootstrap replicates thereof.
}
\description{
Estimate components for sensitivity analysis.
}
\examples{
\dontrun{
data(carnegie_marinov)
y <- "new_empinxavg"
m <- "pwt_openk"
d <- "EV"
z <- "l2CPcol2"
fe <- c("ccode", "year")
out <- estimate_sensitivity(carnegie_marinov, R = 1501,
                     outcome = y, treat = d, inst = z, post.inst = m,
                     fe = fe,
                     seed = 823,
                     parallel = "snow",
                     ncpus = 4)
                      }
}
\references{
Schuessler, J., Glynn, A. N., & Rueda, M.R. (2023). Post-Instrument Bias.

Glynn, A. N., Rueda, M. R., & Schuessler, J. (2023). Post-Instrument Bias in Linear Models. Sociological Methods & Research.
}
