% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hong_data}
\alias{hong_data}
\title{Data from Hong, Park and Yang 2023}
\format{
An object of class \code{data.frame} with 1460 rows and 264 columns.
}
\source{
\url{https://doi.org/10.7910/DVN/MWS8QF}
}
\usage{
hong_data
}
\description{
A subset of data from Hong, Ji Yeon, Sunkyoung Park and Hyunjoo Yang. 2023.
“In strongman we trust: The political legacy of the new village movement in
South Korea.” American Journal of Political Science 67(4):850–866.
}
\keyword{datasets}
