% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{out_cm}
\alias{out_cm}
\title{Estimated components for sensitivity analysis for Carnegie/Marinov 2017}
\format{
An object of class \code{boot} of length 11.
}
\usage{
out_cm
}
\description{
Estimated components for sensitivity analysis for Carnegie/Marinov 2017
}
\examples{
\dontrun{
y <- "new_empinxavg"
m <- "pwt_openk"
d <- "EV"
z <- "l2CPcol2"
fe <- c("ccode", "year")
out <- estimate_sensitivity(carnegie_marinov, R = 1501,
                     outcome = y, treat = d, inst = z, post.inst = m,
                     fe = fe,
                     seed = 823,
                     parallel = "snow",
                     ncpus = 4)
                      }
}
\keyword{datasets}
