% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pib-package.R
\docType{package}
\name{pib-package}
\alias{pib}
\alias{pib-package}
\title{pib: Sensitivity Analysis for Post-Instrument Bias}
\description{
Tools to assess the sensitivity of estimates when the instrument is thought to affect the outcome through a measured variable.
}
\author{
\strong{Maintainer}: Julian Schuessler \email{polsci@julianschuessler.net} (\href{https://orcid.org/0000-0002-8604-7699}{ORCID})

}
\keyword{internal}
