% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pib_bounds.R
\name{plot_pib_bounds}
\alias{plot_pib_bounds}
\title{Plot estimated bounds for sensitivity analysis with respect to post-instrument covariates.}
\usage{
plot_pib_bounds(
  bounds,
  whichCI = "percentile",
  annotate.sigma = "minmax",
  digits = 2,
  set.ylim = T,
  set.xlim = T,
  ylab = "Effect Estimate",
  xlab = expression(paste("E[", gamma[i], "]")),
  ylim = NULL,
  xlim = NULL,
  horizontal.lty = "dashed",
  ...
)
}
\arguments{
\item{bounds}{An object from \code{\link[=post.instrument.bounds]{post.instrument.bounds()}}}

\item{whichCI}{Determines the type of boostrap confidence interval
to be plotted. Either \code{"percentile"} (the default) or \code{"basic"}}

\item{annotate.sigma}{Should the confidence intervals have annotated
values of sigma? Default is \code{"minmax"} (annotating CI for
lowest and highest values of sigma). Alternatives can be
\code{"none"} or \code{"all"}}

\item{digits}{When annotating \code{sigma}, this controls the rounding of
the annotated values. Default is 2.}

\item{set.ylim}{Logical. Should limit of Y-axis be automatically
determined? If \code{FALSE}, argument \code{ylim} has to be set.
Default is \code{TRUE}}

\item{set.xlim}{Logical. Should limit of X-axis be automatically
determined? If \code{FALSE}, argument \code{xlim} has to be set.
Default is \code{TRUE}}

\item{ylab, xlab, ...}{Label of Y- and X-axis, and further graphical parameters.}

\item{xlim, ylim}{If \code{set.ylim = F} or \code{set.ylim = F}, sets limits of axes.}

\item{horizontal.lty}{Controls appearance of horizontal line at 0. Default is \code{"dashed"}.}
}
\description{
Plot estimated bounds for sensitivity analysis with respect to post-instrument covariates.
}
