% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_instrument_bounds.R
\name{post_instrument_bounds}
\alias{post_instrument_bounds}
\title{Compute bounds from estimated components.}
\usage{
post_instrument_bounds(
  gamma,
  sigma = 0,
  n.sigma = 5,
  estimates,
  df,
  outcome,
  post.inst,
  whichCI = "percentile",
  remove.outliers = T,
  conf = 0.95
)
}
\arguments{
\item{gamma}{Integer or vector of integers giving lowest and highest
value for mean causal effect of post-instrument variable \code{M} on outcome \code{Y}}

\item{sigma}{Integer or vector of integers giving lowest and highest
value for variance of causal effect of M on Y. Must
be non-negative. Defaults to zero.}

\item{n.sigma}{Integer: Number of distinct values of \code{sigma} (within bounds provided by
\code{sigma} argument) for which to compute bounds. Defaults to 5.}

\item{estimates}{An object from calling \code{\link[=estimate_sensitivity]{estimate_sensitivity()}} that
contains a vector of point estimates \code{t} as
well as bootstrap replicates \code{t0}}

\item{df}{Data frame}

\item{outcome}{Name of dependent variables of interest \code{Y}, a string}

\item{post.inst}{Name of post-instrument variable \code{M}, a string}

\item{whichCI}{Determines the type of boostrap confidence interval:
\code{percentile} (the default), \code{basic}, or \code{both}}

\item{remove.outliers}{Logical. If \code{TRUE} (the default), outliers from bootstrapped
sampling distribution of the 2SLS regressions for \code{Y} and \code{M} are removed.
The resulting sampling distribution should better
approximate results from standard robust standard errors.}

\item{conf}{The confidence level of the confidence intervals. Defaults
to \code{0.95}}
}
\value{
A matrix with values of \code{gamma}, \code{sigma}, point estimates or bounds
of treatment effect, and confidence intervals for those point estimates
or bounds.
}
\description{
Compute bounds from estimated components.
}
\details{
The function assumes that the instrument \code{Z}, treatment \code{D}, and post-instrument
variable \code{M} are binary (in which case there is only the sensitivity parameter
\code{gamma} and \code{sigma} can be left at its default value zero), or that
\code{z} is binary and \code{D} and \code{M} continuous, in which case both \code{gamma} and
\code{sigma} are sensitivity parameters.
}
\examples{
\dontrun{
gamma <- c(-0.02, 0.02)
sigma <- c(0, 0.002)
bounds <- post.instrument.bounds(gamma = gamma, sigma = sigma,
n.sigma = 2, estimates = out)}
}
\references{
Schuessler, J., Glynn, A. N., & Rueda, M.R. (2023). Post-Instrument Bias.

Glynn, A. N., Rueda, M. R., & Schuessler, J. (2023). Post-Instrument Bias in Linear Models. Sociological Methods & Research.
}
