% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spenkuch_tillmann}
\alias{spenkuch_tillmann}
\title{Data from Spenkuch/Tillmann 2017}
\format{
An object of class \code{data.frame} with 982 rows and 56 columns.
}
\source{
\url{https://doi.org/10.7910/dvn/j2fa5r}
}
\usage{
spenkuch_tillmann
}
\description{
A subset of data from Spenkuch, Jörg L., and Philipp Tillmann.
"Elite influence? Religion and the electoral success of the Nazis." American Journal of Political Science 62.1 (2018): 19-36.
}
\keyword{datasets}
